fix(keycloak): restore direct RabbitMQ publishing
