+++ packages/keycloak-extensions/custom-event-listener/pom.xml
@@ -73,12 +73,6 @@ SPDX-License-Identifier: AGPL-3.0-only
 			<artifactId>keycloak-services</artifactId>
 		</dependency>
 		<dependency>
-			<groupId>org.keycloak</groupId>
-			<artifactId>keycloak-model-jpa</artifactId>
-			<version>${keycloak.version}</version>
-			<scope>provided</scope>
-		</dependency>
-		<dependency>
 			<groupId>org.projectlombok</groupId>
 			<artifactId>lombok</artifactId>
 		</dependency>
