+++ packages/keycloak-extensions/custom-event-listener/src/main/java/sequent/keycloak/custom_event_listener/RabbitMqEventPublisher.java
@@ -5,21 +5,27 @@
 package sequent.keycloak.custom_event_listener;
 
 import com.rabbitmq.client.AMQP;
+import com.rabbitmq.client.Channel;
+import com.rabbitmq.client.Connection;
 import com.rabbitmq.client.ConnectionFactory;
-import java.util.Map;
-import java.util.Objects;
+import java.io.UnsupportedEncodingException;
+import java.net.URLEncoder;
+import java.nio.charset.StandardCharsets;
 import java.util.Optional;
-import org.keycloak.models.KeycloakSession;
-import org.keycloak.models.KeycloakSessionFactory;
+import java.util.function.Supplier;
+import lombok.extern.jbosslog.JBossLog;
 
+@JBossLog
 class RabbitMqEventPublisher implements AutoCloseable {
 
   private static final String DEFAULT_QUEUE_NAME = "electoral_log_event_queue";
-  private static final String TASK_HEADER = "task";
 
+  private final String amqpUri;
   private final String queueName;
-  private final AuditOutboxStore outboxStore;
-  private final RabbitMqOutboxWorker worker;
+  private final Supplier<ConnectionFactory> connectionFactorySupplier;
+
+  private Connection rabbitConnection;
+  private Channel rabbitChannel;
 
   static RabbitMqEventPublisher fromEnvironment() {
     String envSlug = System.getenv("ENV_SLUG");
@@ -29,41 +35,117 @@ class RabbitMqEventPublisher implements AutoCloseable {
         envSlug == null || envSlug.trim().isEmpty()
             ? baseQueueName
             : envSlug.trim() + "_" + baseQueueName;
-    AuditOutboxStore outboxStore = new JpaAuditOutboxStore();
-    RabbitMqOutboxWorker worker =
-        new RabbitMqOutboxWorker(System.getenv("AMQP_ADDR"), outboxStore, ConnectionFactory::new);
-    return new RabbitMqEventPublisher(queueName, outboxStore, worker);
+    return new RabbitMqEventPublisher(
+        System.getenv("AMQP_ADDR"), queueName, ConnectionFactory::new);
   }
 
   RabbitMqEventPublisher(
-      String queueName, AuditOutboxStore outboxStore, RabbitMqOutboxWorker worker) {
-    this.queueName = Objects.requireNonNull(queueName, "queueName");
-    this.outboxStore = Objects.requireNonNull(outboxStore, "outboxStore");
-    this.worker = Objects.requireNonNull(worker, "worker");
+      String amqpUri, String queueName, Supplier<ConnectionFactory> connectionFactorySupplier) {
+    this.amqpUri = encodeAmqpUri(amqpUri);
+    this.queueName = queueName;
+    this.connectionFactorySupplier = connectionFactorySupplier;
   }
 
-  void publish(KeycloakSession session, AMQP.BasicProperties properties, byte[] body) {
-    String correlationId = requireValue(properties.getCorrelationId(), "correlation ID");
-    Map<String, Object> headers = properties.getHeaders();
-    Object taskHeader = headers == null ? null : headers.get(TASK_HEADER);
-    String taskName = requireValue(taskHeader == null ? null : taskHeader.toString(), "task name");
-    outboxStore.enqueue(
-        session, new RabbitMqOutboxMessage(correlationId, queueName, taskName, body, 0, null));
+  synchronized void publish(AMQP.BasicProperties properties, byte[] body) throws Exception {
+    try {
+      getRabbitChannel().basicPublish("", queueName, properties, body);
+    } catch (Exception exception) {
+      closeRabbitMqResources();
+      throw exception;
+    }
   }
 
-  void start(KeycloakSessionFactory sessionFactory) {
-    worker.start(sessionFactory);
+  private Channel getRabbitChannel() throws Exception {
+    if (rabbitConnection == null || !rabbitConnection.isOpen()) {
+      initializeRabbitMqConnection();
+    } else if (rabbitChannel == null || !rabbitChannel.isOpen()) {
+      closeRabbitChannel();
+      rabbitChannel = rabbitConnection.createChannel();
+      declareQueue();
+    }
+    return rabbitChannel;
+  }
+
+  private void initializeRabbitMqConnection() throws Exception {
+    closeRabbitMqResources();
+
+    ConnectionFactory rabbitFactory = connectionFactorySupplier.get();
+    rabbitFactory.setAutomaticRecoveryEnabled(false);
+    rabbitFactory.setTopologyRecoveryEnabled(false);
+    rabbitFactory.setUri(amqpUri);
+
+    try {
+      rabbitConnection = rabbitFactory.newConnection();
+      rabbitChannel = rabbitConnection.createChannel();
+      declareQueue();
+      log.info("RabbitMQ connection and channel initialized.");
+    } catch (Exception exception) {
+      closeRabbitMqResources();
+      throw exception;
+    }
+  }
+
+  private void declareQueue() throws Exception {
+    rabbitChannel.queueDeclare(queueName, true, false, false, null);
   }
 
   @Override
-  public void close() {
-    worker.close();
+  public synchronized void close() {
+    closeRabbitMqResources();
   }
 
-  private static String requireValue(String value, String name) {
-    if (value == null || value.isBlank()) {
-      throw new AuditEventPersistenceException("Audit event " + name + " is missing");
+  private void closeRabbitMqResources() {
+    closeRabbitChannel();
+
+    Connection connectionToClose = rabbitConnection;
+    rabbitConnection = null;
+    if (connectionToClose == null || !connectionToClose.isOpen()) {
+      return;
     }
-    return value;
+    try {
+      connectionToClose.close();
+    } catch (Exception exception) {
+      log.warn("Error closing RabbitMQ connection", exception);
+    }
+  }
+
+  private void closeRabbitChannel() {
+    Channel channelToClose = rabbitChannel;
+    rabbitChannel = null;
+    if (channelToClose == null || !channelToClose.isOpen()) {
+      return;
+    }
+    try {
+      channelToClose.close();
+    } catch (Exception exception) {
+      log.warn("Error closing RabbitMQ channel", exception);
+    }
+  }
+
+  private static String encodeAmqpUri(String rawAmqpUri) {
+    if (rawAmqpUri == null || !rawAmqpUri.startsWith("amqp://")) {
+      return rawAmqpUri;
+    }
+
+    try {
+      String afterScheme = rawAmqpUri.substring("amqp://".length());
+      int atIndex = afterScheme.indexOf('@');
+      if (atIndex == -1) {
+        return rawAmqpUri;
+      }
+
+      String userInfo = afterScheme.substring(0, atIndex);
+      String afterUserInfo = afterScheme.substring(atIndex + 1);
+      String[] userPass = userInfo.split(":", 2);
+      String user = userPass[0];
+      String password = userPass.length > 1 ? userPass[1] : "";
+      String encodedUser = URLEncoder.encode(user, StandardCharsets.UTF_8.name());
+      String encodedPassword = URLEncoder.encode(password, StandardCharsets.UTF_8.name());
+      String encodedUserInfo = encodedUser + (password.isEmpty() ? "" : ":" + encodedPassword);
+
+      return "amqp://" + encodedUserInfo + "@" + afterUserInfo;
+    } catch (UnsupportedEncodingException exception) {
+      throw new IllegalStateException("UTF-8 encoding not supported", exception);
+    }
   }
 }
