--- packages/keycloak-extensions/custom-event-listener/src/main/resources/META-INF/audit-outbox-changelog.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-
-<!--
-SPDX-FileCopyrightText: 2026 Sequent Tech Inc <legal@sequentech.io>
-
-SPDX-License-Identifier: AGPL-3.0-only
--->
-
-<databaseChangeLog
-    xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.1.xsd">
-  <changeSet author="sequent" id="meta-12597-audit-outbox-1">
-    <createTable tableName="SEQUENT_AUDIT_OUTBOX">
-      <column name="ID" type="VARCHAR(36)">
-        <constraints nullable="false" primaryKey="true" primaryKeyName="PK_SEQ_AUDIT_OUTBOX"/>
-      </column>
-      <column name="QUEUE_NAME" type="VARCHAR(255)">
-        <constraints nullable="false"/>
-      </column>
-      <column name="TASK_NAME" type="VARCHAR(255)">
-        <constraints nullable="false"/>
-      </column>
-      <column name="PAYLOAD" type="CLOB">
-        <constraints nullable="false"/>
-      </column>
-      <column name="CREATED_AT" type="BIGINT">
-        <constraints nullable="false"/>
-      </column>
-      <column name="AVAILABLE_AT" type="BIGINT">
-        <constraints nullable="false"/>
-      </column>
-      <column name="CLAIMED_UNTIL" type="BIGINT"/>
-      <column name="CLAIM_TOKEN" type="VARCHAR(36)"/>
-      <column name="ATTEMPTS" type="INT" defaultValueNumeric="0">
-        <constraints nullable="false"/>
-      </column>
-      <column name="LAST_ERROR" type="VARCHAR(2048)"/>
-    </createTable>
-    <createIndex indexName="IDX_SEQ_AUDIT_OUTBOX_READY" tableName="SEQUENT_AUDIT_OUTBOX">
-      <column name="AVAILABLE_AT"/>
-      <column name="CLAIMED_UNTIL"/>
-      <column name="CREATED_AT"/>
-    </createIndex>
-  </changeSet>
-</databaseChangeLog>
